renamed get_docsum to geo_data
